Update inconvenient relative imports + move config into 'js'
