+++ amundsen_application/static/js/components/BrowsePage/index.tsx
@@ -5,14 +5,14 @@ import { bindActionCreators } from 'redux';
 
 import './styles.scss';
 
-import AppConfig from '../../../config/config';
-import LoadingSpinner from '../common/LoadingSpinner';
-import TagInfo from "../Tags/TagInfo";
-import { Tag } from "../Tags/types";
+import AppConfig from 'config/config';
+import LoadingSpinner from 'components/common/LoadingSpinner';
+import TagInfo from 'components/Tags/TagInfo';
+import { Tag } from 'components/Tags/types';
 
-import { GlobalState } from "../../ducks/rootReducer";
-import { getAllTags } from '../../ducks/allTags/reducer';
-import { GetAllTagsRequest } from "../../ducks/allTags/types";
+import { GlobalState } from 'ducks/rootReducer';
+import { getAllTags } from 'ducks/allTags/reducer';
+import { GetAllTagsRequest } from 'ducks/allTags/types';
 
 export interface StateFromProps {
   allTags: Tag[];
