+++ amundsen_application/static/js/components/Footer/index.tsx
@@ -4,9 +4,9 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 // TODO: Use css-modules instead of 'import'
 import './styles.scss';
-import { GlobalState } from "../../ducks/rootReducer";
-import { getLastIndexed } from "../../ducks/tableMetadata/reducer";
-import { GetLastIndexedRequest } from "../../ducks/tableMetadata/types";
+import { GlobalState } from 'ducks/rootReducer';
+import { getLastIndexed } from 'ducks/tableMetadata/reducer';
+import { GetLastIndexedRequest } from 'ducks/tableMetadata/types';
 
 // Props
 interface StateFromProps {
