+++ amundsen_application/static/js/components/NavBar/index.tsx
@@ -4,10 +4,10 @@ import { Link, NavLink, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import AppConfig from '../../../config/config';
-import { GlobalState } from "../../ducks/rootReducer";
-import { getLoggedInUser } from "../../ducks/user/reducer";
-import { LoggedInUser, GetLoggedInUserRequest } from "../../ducks/user/types";
+import AppConfig from 'config/config';
+import { GlobalState } from 'ducks/rootReducer';
+import { getLoggedInUser } from 'ducks/user/reducer';
+import { LoggedInUser, GetLoggedInUserRequest } from 'ducks/user/types';
 
 import './styles.scss';
