+++ amundsen_application/static/js/config/config-custom.ts
@@ -1,6 +1,6 @@
 // This file should be used to add new config variables or overwrite defaults from config-default.ts
 
-import { AppConfigCustom } from './config.types';
+import { AppConfigCustom } from './config-types';
 
 const configCustom: AppConfigCustom = {
   browse: {
