+++ amundsen_application/static/js/ducks/allTags/types.ts
@@ -1,4 +1,4 @@
-import { Tag } from '../../components/Tags/types';
+import { Tag } from 'components/Tags/types';
 export { Tag };
 
 /* API */
