+++ amundsen_application/static/js/ducks/popularTables/types.ts
@@ -1,4 +1,4 @@
-import { TableResource } from '../../components/common/ResourceListItem/types';
+import { TableResource } from 'components/common/ResourceListItem/types';
 export { TableResource };
 
 /* API */
