+++ amundsen_application/static/js/ducks/tableMetadata/types.ts
@@ -1,6 +1,6 @@
-import { PreviewData, PreviewQueryParams, TableMetadata, User } from '../../components/TableDetail/types';
-import { UpdateTagData, Tag } from '../../components/Tags/types';
-import { UpdateMethod } from '../../components/TableDetail/OwnerEditor/types';
+import { PreviewData, PreviewQueryParams, TableMetadata, User } from 'components/TableDetail/types';
+import { UpdateTagData, Tag } from 'components/Tags/types';
+import { UpdateMethod } from 'components/TableDetail/OwnerEditor/types';
 export { PreviewData, PreviewQueryParams, TableMetadata, Tag, User, UpdateMethod, UpdateTagData };
 
 type MessageResponse = { msg: string };
